fix: move SampleList tests to recorded playback, fix line-number comments
- SampleList and SampleList_JSON were labeled Tier 0 (offline) but made
  live HTTPS calls to aka.ms/foundry-agents-samples. Move them to Tier 1
  with recording sessions so they replay from cassette deterministically.
- Remove hardcoded source line numbers from comments (they drift).
- Tier 0 now has 4 truly offline tests; Tier 1 has 5 recorded tests.
