+++ cli/azd/test/functional/ai_agents/ai_agent_test.go
@@ -6,7 +6,6 @@
 package ai_agents_test
 
 import (
-	"encoding/json"
 	"os"
 	"testing"
 
@@ -71,40 +70,8 @@ func Test_AIAgent_Init_NoPrompt_MissingFlags(t *testing.T) {
 		"should fail with clear validation error, not crash")
 }
 
-// Test_AIAgent_SampleList verifies sample list returns results.
-func Test_AIAgent_SampleList(t *testing.T) {
-	t.Parallel()
-	ctx, cancel := newTestContext(t)
-	defer cancel()
-
-	dir := tempDirWithDiagnostics(t)
-	cli := azdcli.NewCLI(t)
-	cli.WorkingDirectory = dir
-	cli.Env = append(cli.Env, os.Environ()...)
-
-	result, err := cli.RunCommand(ctx, "ai", "agent", "sample", "list")
-	require.NoError(t, err, "stdout=%s, stderr=%s", result.Stdout, result.Stderr)
-	require.Greater(t, len(result.Stdout), 50, "sample list output too short")
-}
-
-// Test_AIAgent_SampleList_JSON verifies sample list --output json returns valid JSON array.
-func Test_AIAgent_SampleList_JSON(t *testing.T) {
-	t.Parallel()
-	ctx, cancel := newTestContext(t)
-	defer cancel()
-
-	dir := tempDirWithDiagnostics(t)
-	cli := azdcli.NewCLI(t)
-	cli.WorkingDirectory = dir
-	cli.Env = append(cli.Env, os.Environ()...)
-
-	result, err := cli.RunCommand(ctx, "ai", "agent", "sample", "list", "--output", "json")
-	require.NoError(t, err, "stdout=%s, stderr=%s", result.Stdout, result.Stderr)
-
-	var output map[string]json.RawMessage
-	require.NoError(t, json.Unmarshal([]byte(result.Stdout), &output), "output is not valid JSON: %s", result.Stdout)
-	require.Contains(t, output, "templates", "expected 'templates' key in JSON output")
-}
+// NOTE: SampleList tests moved to ai_agent_recording_test.go (Tier 1) to avoid
+// live network dependency — sample list fetches from aka.ms/foundry-agents-samples.
 
 // Test_AIAgent_Doctor_Help verifies doctor --help shows usage.
 func Test_AIAgent_Doctor_Help(t *testing.T) {
